Split board IDs list into separate file.
